Fix os_network for non-admin user operations
This change alters os_network to support non-admin level operations.
Creating a provider network and specifying a different tenant can
be done only by an admin, and thus belongs in a different module.

This also changes the module to not use the neutron client directly,
but use new shade APIs instead.